+++ src/components/nav/Nav.js
@@ -1,5 +1,6 @@
 import './Nav.scss';
 import { Link } from 'react-router-dom';
+import NavItems from './NavItem';
 
 function Nav() {
   return (
@@ -16,7 +17,7 @@ function Nav() {
             </Link>
           </h1>
           <nav className="wesullocNav">
-            <ul className="navList">
+            {/* <ul className="navList">
               {NAV_LIST.map(({ id, link, title }) => (
                 <li key={id} className="navItems">
                   <Link to={link} className="navTitle">
@@ -24,7 +25,8 @@ function Nav() {
                   </Link>
                 </li>
               ))}
-            </ul>
+            </ul> */}
+            <NavItems />
           </nav>
         </div>
 
@@ -53,21 +55,22 @@ function Nav() {
     </header>
   );
 }
-const NAV_LIST = [
-  {
-    id: 1,
-    link: '/',
-    title: '제품',
-  },
-  {
-    id: 2,
-    link: '/',
-    title: '선물추천',
-  },
-  {
-    id: 3,
-    link: '/',
-    title: '다다일상',
-  },
-];
+// const NAV_LIST = [
+//   {
+//     id: 1,
+//     link: '/',
+//     title: '제품',
+//   },
+//   {
+//     id: 2,
+//     link: '/',
+//     title: '선물추천',
+//   },
+//   {
+//     id: 3,
+//     link: '/',
+//     title: '다다일상',
+//   },
+// ];
+
 export default Nav;
