+++ src/components/nav/Nav.scss
@@ -48,6 +48,7 @@
         }
       }
     }
+
     .rightBox {
       display: flex;
       color: $bgWhite;
